DDA: do the first step in dda_start().
The problem is that when one movement ends and another one starts,
both, dda_step() and dda_start() have to be executed. The idea
here, together with the previous commit, is to massage the code in
a way that it executes only one of the between to steps, no matter
wether it's a step in midst of a movement of the step between two
movements.

Performance improvement is moderate. Not as slow as what we had
before this optimisation series started, but also these single
slow steps between movements not going away.

  ATmega sizes               '168   '328(P)   '644(P)     '1280
  Program:  19868 bytes      139%       65%       32%       16%
     Data:   2255 bytes      221%      111%       56%       28%
   EEPROM:     32 bytes        4%        2%        2%        1%

  short-moves.gcode statistics:
  LED on occurences: 907.
  LED on time minimum: 241 clock cycles.
  LED on time maximum: 648 clock cycles.
  LED on time average: 317.47 clock cycles.

  smooth-curves.gcode statistics:
  LED on occurences: 23681.
  LED on time minimum: 241 clock cycles.
  LED on time maximum: 653 clock cycles.
  LED on time average: 351.711 clock cycles.

  triangle-odd.gcode statistics:
  LED on occurences: 1639.
  LED on time minimum: 241 clock cycles.
  LED on time maximum: 646 clock cycles.
  LED on time average: 334.508 clock cycles.